requirements.txt
Django~=4.0
djangorestframework==3.13.1
whitenoise
gunicorn==20.1.0
psycopg2-binary
django-celery-beat==2.3.0
django-environ==0.9.0
django-extensions
redis==4.3.4
aiohttp==3.8.3
celery==5.2.7
pytest-django
pytest-cov
django-filter
drf-spectacular==0.21.2
requests
